# java-bootcamp-2018-tdd/Market_1.0_swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "defaultTitle",
    "description" : "defaultDescription",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ],
  "paths" : {
  "/user/update/lastName" : {
      "put" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "lastName" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "nickName" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"nickName\":\"User2NickName\",\n    \"password\":\"User2Password\",\n    \"lastName\":\"lastNameExample\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector"
          }
        }
      }
    },
...
    "/user/update/password" : {
      "put" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstName" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "nickName" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"nickName\":\"User2NickName\",\n    \"password\":\"User2Password\",\n    \"firstName\":\"passwordExample\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector"
          }
        }
      }
    },
    "/user/create" : {
      "post" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstName" : {
                    "type" : "string"
                  },
                  "lastName" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "nickName" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"nickName\":\"User8NickName\",\n\t\"firstName\":\"User8FirstName\",\n\t\"lastName\":\"User8LastName\",\n\t\"password\":\"User8Password\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector"
          }
        }
      }
    },
...
    "/user/nickName" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "parameters" : [ {
          "name" : "nickName",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "example" : "User7NickName"
        } ],
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json;charset=UTF-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        }
      }
    },
...
    "/user/name" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "parameters" : [ {
          "name" : "firstName",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "example" : "User2FirstName"
        }, {
          "name" : "lastName",
          "in" : "query",
          "schema" : {
            "type" : "string"
          },
          "example" : "User2LastName"
        } ],
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json;charset=UTF-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        }
      }
    },
...
    "/user/update/firstName" : {
      "put" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "firstName" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "nickName" : {
                    "type" : "string"
                  }
                }
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"nickName\":\"User2NickName\",\n    \"password\":\"User2Password\",\n    \"firstName\":\"firstNameExample\"\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector"
          }
        }
      }
    },
...
  }
}